+++ src/context/github/githubState.js
@@ -57,7 +57,21 @@ const GithubState = props => {
     }
 
     // Get Repos
+    const getUserRepos = async (username) => {
+        setLoading(true);
 
+        let url = `https://api.github.com/users/${username}/repos?` +
+            `per_page=5&sort=created:asc&` +
+            `client_id=${process.env.REACT_APP_GITHUB_CLIENT_ID}&` +
+            `client_secret=${process.env.REACT_APP_GITHUB_CLIENT_SECRET}`;
+
+        await axios.get(url).then(res => {
+            dispatch({
+                type: GET_REPOS,
+                payload: res.data
+            });
+        });
+    }
     // Clear Users
     const clearUsers = () => {
         dispatch({
@@ -80,7 +94,8 @@ const GithubState = props => {
             loading: state.loading,
             searchUsers,
             clearUsers,
-            getUser
+            getUser,
+            getUserRepos
         }}
     >
         {props.children}
